test: adapt test helpers to uint64 epoch API
- PaymentsTestHelpers:
   - getAccountData() now unpacks lockupLastSettledAt as uint64  and return Payments.Account matching new signature
  - assertAccountState expectedLastSettled parameter changed to uint64
  - setupRailWithParameters accepts lockupPeriod as uint64
- RailSettlementHelpers:
  - SettlementResult.settledUpto field changed to uint64
  - settleRailAndVerify and terminateAndSettleRail signatures updated to use uint64
  - Added explicit `uint64(block.number)` casts where needed

+++ test/helpers/PaymentsTestHelpers.sol
@@ -104,7 +104,7 @@ contract PaymentsTestHelpers is Test, BaseTestHelper {
             uint256 funds,
             uint256 lockupCurrent,
             uint256 lockupRate,
-            uint256 lockupLastSettledAt
+            uint64 lockupLastSettledAt
         ) = payments.accounts(token, user);
 
         return
@@ -340,7 +340,7 @@ contract PaymentsTestHelpers is Test, BaseTestHelper {
         address to,
         address railOperator,
         uint256 paymentRate,
-        uint256 lockupPeriod,
+        uint64 lockupPeriod,
         uint256 lockupFixed,
         address arbiter
     ) public returns (uint256 railId) {
@@ -549,7 +549,7 @@ contract PaymentsTestHelpers is Test, BaseTestHelper {
         uint256 expectedFunds,
         uint256 expectedLockup,
         uint256 expectedRate,
-        uint256 expectedLastSettled
+        uint64 expectedLastSettled
     ) public view {
         Payments.Account memory account = getAccountData(user);
         assertEq(account.funds, expectedFunds, "Account funds incorrect");
